requirements.txt
streamlit
openai
tiktoken
python-dotenv
cryptography
supabase
email-validator